refactor: 重构 NullableTime 值类型与响应数据结构
变更点：
- 将 NullableTime 从指针类型改为值类型，消除空指针风险
- 提取 ResBase 公共响应结构体，简化各响应类型定义
- 移除废弃的 infrastructure/utils 包（timeParser.go、nullableTime.go），统一使用 domain/types/nullableTime.go
- 重命名控制器和类型文件（auth→identity, common→nullableString, entity-base→entityBase），统一命名风格
- 新增 GetPomodoroRes/CreateTaskCommentRes 类型别名，清理重复类型定义
- 在路由初始化中增加 Debug/Release 模式切换和信任代理 IP 配置
- 将 UpdateTaskReq 中的 NullableString 类型简化为 *string 指针类型
- 统一代码格式，将长链式调用按方法分行，提升可读性

其他:
- 调整 pomodoro Duration 字段类型从 uint8 改为 uint16 以支持更大时长
- 调整 state 和 priority 查询类型从 []uint8 改为 []int，修复 SQL 兼容性

+++ domain/pomodoro/entities/pomodoro.go
@@ -18,6 +18,6 @@ type Pomodoro struct {
 	Description string
 	StartAt     time.Time
 	EndAt       time.Time
-	Duration    uint8
+	Duration    uint16
 	Note        string
 }
